toy2/src/main/java/toy2/toy2/service/member/MemberServiceImpl.java
package toy2.toy2.service.member;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Component;
import toy2.toy2.domain.Member;
import toy2.toy2.repository.MemberRepository;

@Component
public class MemberServiceImpl implements MemberService{

    private final MemberRepository memberRepository;

    @Autowired
    public MemberServiceImpl(MemberRepository memberRepository) {
        this.memberRepository = memberRepository;
    }

    @Override
    public void join(Member member) {
        memberRepository.save(member);
    }

    @Override
    public void remove(Member member) {
        memberRepository.remove(member);
    }

    @Override
    public Member findMember(Long memberId) {
        return memberRepository.findById(memberId);
    }

    @Override
    public int sizeMember() {
        return memberRepository.printSize();
    }
}
